refactor(file-memory-store):Enforce the maxFiles guard against the key listing inside readAllFiles,   before any content is read, so an oversized store is rejected without   first pulling its whole corpus into memory, Extract mapWithConcurrency and STORAGE_READ_CONCURRENCY into a neutral   concurrency module so search() no longer imports from the   consolidation-specific execute module, and make comments more concise
