+++ strands-ts/src/vended-memory-stores/file-memory-store/consolidation/execute.ts
@@ -14,49 +14,20 @@ import type { ConsolidateOperation } from '../types.js'
 import type { ConsolidationAction, ConsolidationPlan } from './plan.js'
 import { ConsolidationError } from '../../../errors.js'
 import { logger } from '../../../logging/logger.js'
+import { mapWithConcurrency, STORAGE_READ_CONCURRENCY } from '../concurrency.js'
 
 const encoder = new TextEncoder()
 const decoder = new TextDecoder()
 
 /**
  * Path (within the store's namespace) reserved for the consolidation audit log. Owned here because
- * this module both writes it ({@link recordChangelog}) and excludes it from the working set
- * ({@link readAllFiles}); {@link FileMemoryStore} and the plan validator import it to keep it out of
- * search results and off every plan's write targets.
+ * this module writes it ({@link recordChangelog}) and excludes it from the working set
+ * ({@link readAllFiles}); {@link FileMemoryStore} and the validator import it to keep it out of
+ * search results and off plan write targets.
  *
  * @internal
  */
 export const CONSOLIDATION_CHANGELOG = 'consolidation-changelog.md'
-
-/**
- * Cap on concurrent storage reads when fanning out over the store's keys. An unbounded fan-out can
- * exhaust a backend's connection pool or trip throttling on a large corpus.
- *
- * @internal
- */
-export const STORAGE_READ_CONCURRENCY = 8
-
-/**
- * Map `items` through `fn` running at most `limit` calls concurrently, preserving input order.
- * A worker pool pulls from a shared cursor so a slow item never blocks others in its batch.
- *
- * @internal
- */
-export async function mapWithConcurrency<T, R>(items: T[], limit: number, fn: (item: T) => Promise<R>): Promise<R[]> {
-  const results = new Array<R>(items.length)
-  let cursor = 0
-  const worker = async (): Promise<void> => {
-    while (cursor < items.length) {
-      const index = cursor++
-      // index is bounded by items.length — the item is always present
-      const item = items[index] as T
-      results[index] = await fn(item)
-    }
-  }
-  const workers = Array.from({ length: Math.min(limit, items.length) }, () => worker())
-  await Promise.all(workers)
-  return results
-}
 
 /**
  * A delete that failed during execution, paired with the error the backend raised.
@@ -73,15 +44,25 @@ export interface DeleteFailure {
 /**
  * Read every knowledge file into memory as a `path → content` map.
  *
- * This snapshot is the working set for one consolidation run: it is handed to the planner,
- * the validator, and the executor so they all reason over the same view of the store.
+ * This snapshot is the working set for one run — the planner, validator, and executor all reason
+ * over it. `maxFiles` is checked against the key listing *before* any content is read, so an
+ * oversized store is rejected without first pulling its whole corpus into memory.
+ *
+ * @throws ConsolidationError when the store holds more than `maxFiles` knowledge files
  *
  * @internal
  */
-export async function readAllFiles(storage: Storage): Promise<Map<string, string>> {
+export async function readAllFiles(storage: Storage, maxFiles: number): Promise<Map<string, string>> {
   const files = new Map<string, string>()
   const allKeys = await storage.list('')
   const keysToRead = allKeys.filter((key) => key !== CONSOLIDATION_CHANGELOG)
+
+  // Check the count before the read fan-out; the changelog is excluded above, so it never counts.
+  if (keysToRead.length > maxFiles) {
+    throw new ConsolidationError(
+      `Knowledge store exceeds consolidation file limit: ${keysToRead.length} files (maxFiles: ${maxFiles})`
+    )
+  }
 
   const entries = await mapWithConcurrency(keysToRead, STORAGE_READ_CONCURRENCY, async (key) => {
     const bytes = await storage.read(key)
@@ -97,16 +78,16 @@ export async function readAllFiles(storage: Storage): Promise<Map<string, string
 /**
  * Apply a validated plan to storage deterministically, in two passes:
  *
- * 1. **Writes** — merged content lands before its sources are removed, so an interrupted run leaves
- *    duplicates rather than dropping content that had nowhere else to live. This protects sources,
- *    not overwrite targets: an `update` rewrites in place by design.
- * 2. **Deletes** — best-effort. All are attempted and failures returned, not thrown, so the caller
- *    can still record the changelog for a partial run. A missing key is a no-op, so a delete only
- *    fails on a genuine backend error.
+ * 1. **Writes** — before deletes, so an interrupted run leaves duplicates rather than dropping
+ *    content with nowhere else to live. Protects sources, not overwrite targets: `update` rewrites
+ *    in place by design.
+ * 2. **Deletes** — best-effort: all attempted, failures returned rather than thrown so the caller
+ *    can still record a partial run. A missing key is a no-op, so a delete only fails on a real
+ *    backend error.
  *
  * Not safe for concurrent use — writes land unconditionally, so a file created after this run's
- * snapshot can be overwritten by a plan that never saw it. {@link Storage} offers no conditional
- * write, so nothing here can detect it.
+ * snapshot can be overwritten by a plan that never saw it, and {@link Storage} has no conditional
+ * write to detect it.
  *
  * @returns The paths whose deletes failed, each with the underlying error (empty when all succeed)
  *
@@ -117,8 +98,7 @@ export async function executePlan(
   plan: ConsolidationPlan,
   files: Map<string, string>
 ): Promise<DeleteFailure[]> {
-  // Writes before deletes — merged content lands before sources are removed. Every path is already
-  // lowercased (stored keys by add(), plan paths by extractPlan), so a path is its own storage key.
+  // Writes before deletes — merged content lands before sources are removed.
   for (const action of plan.actions) {
     if (action.action === 'merge') {
       await storage.write(action.target, encoder.encode(action.content))
@@ -152,14 +132,12 @@ export async function executePlan(
 /**
  * Collect every path the plan removes, in action order.
  *
- * Gathering them up front keeps the failure-recording try/catch in one place rather than repeating it
- * per action shape.
+ * Gathering them up front keeps the failure-recording try/catch in one place.
  *
  * @returns The paths to delete, which may repeat when two actions name the same source
  */
 function collectDeletePaths(plan: ConsolidationPlan): string[] {
-  // Appended one at a time, never spread — a plan with a huge sources array would blow the argument
-  // limit
+  // Appended one at a time, never spread — a huge sources array would blow the argument limit
   const pathsToDelete: string[] = []
   for (const action of plan.actions) {
     switch (action.action) {
